Introduce SplitStore and Service to handle BL
