+++ core/src/main/java/bisq/core/account/witness/AccountAgeWitnessStore.java
@@ -19,8 +19,7 @@ package bisq.core.account.witness;
 
 import bisq.network.p2p.storage.P2PDataStorage;
 import bisq.network.p2p.storage.payload.PersistableNetworkPayload;
-
-import bisq.common.proto.persistable.ThreadedPersistableEnvelope;
+import bisq.network.p2p.storage.persistence.SplitStore;
 
 import com.google.protobuf.Message;
 
@@ -39,7 +38,7 @@ import lombok.extern.slf4j.Slf4j;
  * definition and provide a hashMap for the domain access.
  */
 @Slf4j
-public class AccountAgeWitnessStore implements ThreadedPersistableEnvelope {
+public class AccountAgeWitnessStore extends SplitStore {
     @Getter
     private Map<P2PDataStorage.ByteArray, PersistableNetworkPayload> map = new ConcurrentHashMap<>();
