+++ p2p/src/main/java/bisq/network/p2p/storage/persistence/SplitStore.java
@@ -1,0 +1,19 @@
+package bisq.network.p2p.storage.persistence;
+
+import bisq.network.p2p.storage.P2PDataStorage;
+import bisq.network.p2p.storage.payload.PersistableNetworkPayload;
+
+import bisq.common.proto.persistable.ThreadedPersistableEnvelope;
+
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
+import lombok.Getter;
+
+/**
+ *  Goes with {@link SplitStoreService}.
+ */
+public abstract class SplitStore implements ThreadedPersistableEnvelope {
+    @Getter
+    protected Map<P2PDataStorage.ByteArray, PersistableNetworkPayload> map = new ConcurrentHashMap<>();
+}
